♻️ 내 위치 관련 코드 컴포넌트 분리
